Move scatter operators into a separate module
Move Scatter, ScatterElements and ScatterND, along with the
`scatter_elements` and `scatter_nd` functions and their tests, from
`ops::gather` into a new `ops::scatter` module.
